chore: Move App component into standard layout (See #1100 and #1260)
