+++ src/amo/components/App/styles.scss
